Refactor ssh_server.py into modular package
Split monolithic ssh_server.py (~1200 lines) into organized modules:
- ssh_server/__init__.py  - package init, re-exports start_ssh_server, _active_connections
- ssh_server/utils.py     - _fingerprint, _lookup_key, _resolve_and_check, _get_user_group
- ssh_server/audit.py     - _record_session_and_audit, _close_session_and_audit, _audit_session_failed
- ssh_server/terminal.py  - NATSTerminal class
- ssh_server/highscores.py - _load/save/add/format_highscores for snake game
- ssh_server/rate_limiter.py - _RateLimiter class
- ssh_server/handlers.py  - _get_menu_agents, SSHSessionHandler, MenuSessionHandler
- ssh_server/server.py    - SSHAgentServer, start_ssh_server, _active_connections

Management command imports unchanged: from tacticalrmm.ssh_server import start_ssh_server, _active_connections
